+++ packages/blade/src/components/Charts/index.ts
@@ -1,4 +1,3 @@
 // Export LineCharts (includes shared components)
-export * from './LineCharts';
-export * from './BaseChartComponents';
-export * from './BarCharts';
+export * from './LineChart';
+export * from './CommonChartComponents';