+++ packages/blade/src/tokens/theme/bladeTheme.ts
@@ -1070,7 +1070,6 @@ const colors: ColorsWithModes = {
         intense: globalColors.neutral.blueGrayDark.a100,
       },
     },
-    //REVIEW_NOTES: this might change
     chart: {
       background: {
         categorical: {